Add save and view renders for regular price and product section
