+++ packages/js/product-editor/src/blocks/product-fields/regular-price/index.ts
@@ -4,6 +4,7 @@
 import blockConfiguration from './block.json';
 import { Edit } from './edit';
 import { Save } from './save';
+import { View } from './view';
 import { registerProductEditorBlockType } from '../../../utils';
 
 const { name, ...metadata } = blockConfiguration;
@@ -14,6 +15,7 @@ export const settings = {
 	example: {},
 	edit: Edit,
 	save: Save,
+	view: View,
 };
 
 export function init() {
